Create ImportAppendTable
This sub routine allows an easy way to import to a specified table and to run any necessary queries.  Currently the code is written to only accept one query but can easily be modified to accept an array of query names.  Data is deleted from the specified table to import new data.  That was the necessary use case when I created this but can be changed to append to the table being imported to.